add log file and check file

+++ run.py
@@ -3,11 +3,14 @@ import os
 import openpyxl
 from dotenv.main import load_dotenv
 import boto3
+from .log import Logger
 
 
 # Create a client with the MinIO server playground, its access key
 
 load_dotenv()
+
+log = Logger()
 
 client = Minio(
     os.environ['URL_MINIO'],
@@ -38,19 +41,44 @@ def download_file(bucket_name, object_name, local_directory):
             # Download the file from MinIO
             client.fget_object(bucket_name, object_name, local_file_path)
 
-            print(f"File downloaded successfully: {local_file_path}")
+            log.info(f"File downloaded successfully to {local_file_path}")
         else:
-            print(f"Bucket '{bucket_name}' does not exist.")
+            log.error(f"Bucket '{bucket_name}' does not exist")
     except Exception as e:
-        print(f"Error downloading file: {str(e)}")
+        log.error(f"Error downloading file: {str(e)}")
+
+def file_exists_in_local(file_path):
+    return os.path.exists(file_path)
+
+
+# def upload_file_to_s3(bucket_name, file_path, object_name):
+#     try:
+#         s3_client.upload_file(file_path, bucket_name, object_name)
+#         log.info(f"File uploaded successfully to {bucket_name}/{object_name}")
+#     except Exception as e:
+#         log.error(f"Error uploading file: {str(e)}")
+
+def check_object_existence(bucket_name, object_name):
+    try:
+        response = s3_client.head_object(Bucket=bucket_name, Key=object_name)
+        return True
+    except s3_client.exceptions.NoSuchKey:
+        return False
+    except Exception as e:
+        print(f"Error checking object existence: {str(e)}")
+        return False
 
 
 def upload_file_to_s3(bucket_name, file_path, object_name):
     try:
-        s3_client.upload_file(file_path, bucket_name, object_name)
-        print(f"File uploaded successfully to {bucket_name}: {object_name}")
+        if not check_object_existence(bucket_name, object_name):
+            s3_client.upload_file(file_path, bucket_name, object_name)
+            log.info(f"File uploaded successfully to {bucket_name}/{object_name}")
+        else:
+            log.info(f"File already exists in {bucket_name}/{object_name}")
     except Exception as e:
-        print(f"Error uploading file: {str(e)}")
+        log.error(f"Error uploading file: {str(e)}")
+
 
 
 path = os.environ['PATH_FOLDER_EXCEL']
@@ -65,24 +93,34 @@ for i in range(1, m_row + 1):
         bucket_name_minio = os.environ['BUCKET_NAME_MINIO']
         object_name_minio = path_url
         local_file_path = "/opt/sampoerna-minio/static"
-        download_file(bucket_name_minio, object_name_minio, local_file_path)
-        print("Download file from Minio success", path_url)
+
+        minio_file_path = os.path.join(local_file_path, os.path.basename(path_url))
+
+        if not file_exists_in_local(minio_file_path):
+            download_file(bucket_name_minio, object_name_minio, local_file_path)
+            log.info("Download file from MinIO success", path_url)
+        else:
+            log.info("File already exists in local static folder, skipping download", path_url)
     except Exception as e:
-        print(f"Error downloading file: {str(e)}")
+        log.error(f"Error downloading file: {str(e)}")
     
     try:
         bucket_name_s3 = os.environ['S3_BUCKET']
         object_name_s3 = path_url
         file_path = "/opt/sampoerna-minio/static/{}".format(path_url.split("/")[-1])
-        upload_file_to_s3(bucket_name_s3, file_path, object_name_s3)
-        print("Upload file to S3 success", path_url)
+
+        if not check_object_existence(bucket_name_s3, object_name_s3):
+            upload_file_to_s3(bucket_name_s3, file_path, object_name_s3)
+            log.info("Upload file to S3 success", path_url)
+        else:
+            log.info("Object already exists in S3, skipping upload", path_url)
     except Exception as e:
-        print(f"Error uploading file: {str(e)}")
+        log.error(f"Error uploading file: {str(e)}")
     
     try:
         os.remove(file_path)
-        print("Delete file success", path_url)
+        log.info("Delete file success", path_url)
     except Exception as e:
-        print(f"Error deleting file: {str(e)}")
+        log.error(f"Error deleting file: {str(e)}")
 
-print("Done")
+log.info("All process success")
